重构数值类型

+++ Assets/Scripts/Gameplay/GameUnit/IProduceable.cs
@@ -6,7 +6,7 @@ namespace Gameplay.GameUnit
     public interface IProduceable
     {
         void       Produce(SoldierUnitBase unit, PlayerBase player, UnitStatus status);
-        public int CostTime        { get; }
+        public float CostTime        { get; }
         public int CostFood        { get; }
         public int CostWood        { get; }
         public int CostGold        { get; }
